+++ Geometry/ForwardGeometry/src/ZdcHardcodeGeometryLoader.cc
@@ -133,7 +133,7 @@ ZdcHardcodeGeometryLoader::makeCell(const HcalZDCDetId& detId,
   }
   GlobalPoint faceCenter(xfaceCenter, yfaceCenter, zfaceCenter);
 
-  std::vector<float> zz ;
+  std::vector<double> zz ;
   zz.reserve(3) ;
   zz.push_back( dx ) ;
   zz.push_back( dy ) ;
